Making donation link a constant string
Making donation link a constant String value in order to prevent any confusion when referring back to this file after a long period of time. It also helps when another file may need to reference the same String containing the link later in the dev process. It's static so that another file doesn't need to instantiate the BaseActivity class in order to retrieve the value of the String.